# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Altas",
    "version": "2.0.0",
    "description": "API REST pour clients, KYC, comptes bancaires, dépôts, retraits, reçus et transferts."
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/clients": {
      "get": {
        "summary": "Rechercher/lister les clients",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" } },
          { "name": "includeArchived", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": { "200": { "description": "Clients", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Client" } } } } } }
      },
      "post": {
        "summary": "Créer un client",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ClientInput" } } } },
        "responses": { "201": { "description": "Client créé", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Client" } } } } }
      }
    },
...
    "/api/accounts": {
      "get": { "summary": "Lister les comptes", "responses": { "200": { "description": "Comptes", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Account" } } } } } } },
      "post": { "summary": "Ouvrir un compte", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateAccountRequest" } } } }, "responses": { "201": { "description": "Compte ouvert" } } }
    },
...
    "/api/transfers": {
      "post": { "summary": "Transfert interne, externe ou passerelle", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransferRequest" } } } }, "responses": { "201": { "description": "Transfert effectué" }, "409": { "description": "Fonds insuffisants" } } }
    },
...
  },
  "components": {
...
    "schemas": {
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "fullName": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "identityNumber": { "type": "string" },
          "photo": { "type": "string" },
          "kycStatus": { "type": "string", "enum": ["PENDING", "VERIFIED", "REJECTED"] },
          "archived": { "type": "boolean" }
        }
      },
      "ClientInput": {
        "type": "object",
        "required": ["firstName", "lastName"],
        "properties": {
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "identityNumber": { "type": "string" },
          "photo": { "type": "string" },
          "kycStatus": { "type": "string", "enum": ["PENDING", "VERIFIED", "REJECTED"] }
        }
      },
      "Account": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "clientId": { "type": "string" },
          "clientName": { "type": "string" },
          "accountNumber": { "type": "string" },
          "type": { "type": "string", "enum": ["CURRENT", "SAVINGS", "BUSINESS", "JOINT"] },
          "status": { "type": "string", "enum": ["ACTIVE", "CLOSED"] },
          "balance": { "type": "number" }
        }
      },
      "CreateAccountRequest": {
        "type": "object",
        "required": ["clientId"],
        "properties": {
          "clientId": { "type": "string" },
          "type": { "type": "string", "enum": ["CURRENT", "SAVINGS", "BUSINESS", "JOINT"] },
          "initialDeposit": { "type": "number", "minimum": 0 }
        }
      },
...
      "TransferRequest": {
        "type": "object",
        "required": ["fromAccountId", "amount"],
        "properties": {
          "fromAccountId": { "type": "string" },
          "toAccountId": { "type": "string" },
          "transferType": { "type": "string", "enum": ["INTERNAL", "EXTERNAL_BANK", "PAYMENT_GATEWAY"] },
          "amount": { "type": "number", "minimum": 0, "exclusiveMinimum": true },
          "beneficiaryName": { "type": "string" },
          "destinationBank": { "type": "string" },
          "gatewayName": { "type": "string" },
          "gatewayReference": { "type": "string" },
          "description": { "type": "string" }
        }
      },
...
    }
  }
}